requirements.txt
absl-py==1.4.0
cython==3.0.2
datasets==2.14.5	
easyocr==1.7.2
evaluate==0.4.0	
fastapi==0.103.1	
flask==2.3.3	
gensim>=4.3.2
imgaug==0.4.0
jamo==0.4.1
konlpy==0.6.0
korpora==0.2.0
lightning>=2.2.0
nlpaug==1.1.11
nltk==3.7
numpy==1.26.4
opencv-python==4.8.0.76
pandas==2.3.3
pillow==9.2.0
portalocker==2.7.0
protobuf==4.25.3
pyarrow==12.0.1
pycocotools==2.0.7
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.0.2	
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.1.97
spacy==3.7.4
streamlit>=1.32
tensorly==0.8.1
timm==0.9.7	
tokenizers==0.13.3	
torch==2.0.1	
torchdata==0.6.1	
torchinfo==1.8.0	
torchtext==0.15.2	
torchvision==0.15.2	
transformers==4.33.2	
ultralytics==8.0.128